fix(epics-pva-rs): keep fractional EPICS_PVA_CONN_TMO before the 4/3 timeout scale
EPICS_PVA_CONN_TMO was parsed as f64 then truncated to integer seconds
(`.map(|f| f.max(1.0) as u64)`) in conn_timeout_secs_opt, so a configured
2.5 became 2 before any consumer applied pvxs's tmoScale (4/3) and the 2s
floor. A Rust client/server then used 2*4/3=2.667s where pvxs uses
2.5*4/3=3.333s and disconnected earlier than the matching pvxs config.

pvxs keeps the operator value as a double end to end: parse_timeout stores
dest = temp * tmoScale (config.cpp:211-227, tmoScale=4/3 config.cpp:148),
and enforceTimeout clamps <2s -> 2s / unset|invalid -> 40s on the double
(config.cpp:373-391).

Structural fix: conn_timeout_secs() / _opt() now return f64 (the raw
positive-finite double), and the 4/3 scale + floor stay owned by each
effective-timeout owner — removing the truncation by construction rather
than re-widening a u64 downstream.

Anchor: env timeout parser that parse::<f64>() then truncates to integer
seconds (as u64/as u32) before consumers scale.
Sites: conn_timeout_secs_opt (config/env.rs, the `as u64`); consumers that
re-cast the truncated value — heartbeat_interval / heartbeat_timeout
(client_native/server_conn.rs), server idle timeout (server_native/runtime.rs).
Same-defect-fix-now: all of the above (env returns f64; the now-redundant
`as f64` casts dropped at each owner).
Distinct-skip: send_timeout_secs_opt and tls_handshake_timeout_secs_opt
(already return f64, floored, no truncation); beacon_period_opt /
beacon_period_long (Duration::from_secs_f64 — sub-second already preserved).

Tests: conn_tmo_preserves_fractional_seconds (env source: 2.5 -> Some(2.5),
unset -> 30.0); heartbeat_timeout_scales_fractional_conn_tmo_with_floor
(owner: below/at/above the 2s effective floor, and != the truncated 2.667s).
testconfig_port assertion updated for the f64 return type.

PVA-RS-2026-05-28-39

+++ crates/epics-pva-rs/src/server_native/runtime.rs
@@ -363,7 +363,7 @@ impl PvaServerConfig {
         // with a healthy client's second ECHO and disconnect it.
         // Floor at 2s mirrors pvxs `enforceTimeout`.
         if let Some(c) = env::conn_timeout_secs_opt() {
-            let scaled = (c as f64 * 4.0 / 3.0).max(2.0);
+            let scaled = (c * 4.0 / 3.0).max(2.0);
             self.idle_timeout = Duration::from_secs_f64(scaled);
         }
         if let Some(v) = env::server_ignore_addr_list_opt() {
